refactor: abstracted base list item with composition
